Controller e rotas para novo formato das notas
