+++ routes/web.php
@@ -11,7 +11,7 @@ use App\Http\Controllers\VeiculoController;
 use App\Http\Controllers\VeiculosClientesController;
 use App\Http\Controllers\ManutencaoController;
 use App\Http\Controllers\OrdemServicoController;
-use App\Http\Controllers\OrdemServicoItemController;
+use App\Http\Controllers\NotasItemController;
 //use App\Http\Controllers\ChatController;
 
 Route::get('/', function () {
@@ -33,7 +33,7 @@ Route::middleware(['auth', 'check.blocked'])->group(function () {
     //Rotas para administradores.
     Route::middleware(['admin'])->group(function () {
         Route::resource('ordemservicos', OrdemServicoController::class);
-        Route::resource('ordemservicoitem', OrdemServicoItemController::class);
+        Route::resource('notasitem', NotasItemController::class);
         Route::resource('users', UserController::class);
         Route::get('/cliente', [ClienteController::class, 'index'])->name('clientes.index');
         Route::resource('produtovendas', produtoVendaController::class);
